feat: build mode mono/micro
