+++ .vscode/launch.json
@@ -6,7 +6,7 @@
             "id": "debugTarget",
             "description": "Please select the launch api ",
             "options": [
-                "main.go",
+                "main.g",
             ]
         }
     ],
@@ -16,9 +16,8 @@
             "type": "go",
             "request": "launch",
             "mode": "debug",
-            "program": "${workspaceFolder}/${input:debugTarget}",
+            "program": "${workspaceFolder}/main.go",
             "env": {},
-            "args": []
         }
     ]
 }
