+++ go.work
@@ -5,5 +5,4 @@ use (
 	./pkg
 	./services/callanotherapi
 	./services/healthcheck
-	./services/intersection
 )
