+++ services/callanotherapi/adapter/controller/call_another_api_controller.go
@@ -2,6 +2,7 @@ package controller
 
 import (
 	"modular-monolith-boilerplate/pkg/di"
+	"modular-monolith-boilerplate/pkg/restapi"
 	"modular-monolith-boilerplate/services/callanotherapi/usecase"
 
 	"github.com/gin-gonic/gin"
@@ -9,6 +10,15 @@ import (
 
 func init() {
 	di.RegisterBean(NewCallAnotherApiController)
+}
+
+func RegisterRouting() {
+	_ = di.GetContainer().Invoke(
+		func(caa *CallAnotherApiController) {
+			group := restapi.NewGroup("/call-another-api")
+			group.RegisterGET("/call", caa.Call)
+		},
+	)
 }
 
 type CallAnotherApiController struct {
