+++ services/callanotherapi/adapter/repository/micro/healthcheck_repository.go
@@ -1,4 +1,4 @@
-package repository
+package micro
 
 import (
 	"fmt"
@@ -25,7 +25,7 @@ func NewMicroHealthCheckRepository() repository.HealthCheckRepository {
 
 func (hcr *MicroHealthCheckRepository) Ping(c *gin.Context) (string, error) {
 	message := ""
-	resp, err := hcr.restClient.CallGet("localhost:8080/health-check/ping", message)
+	resp, err := hcr.restClient.CallGet("/health-check/ping", message)
 	if err != nil {
 		return "", err
 	}
